Extract DocumentationEngine from CoreReplApp.Documentation partial
Move all documentation model generation logic into a standalone
DocumentationEngine class. CoreReplApp.Documentation.cs becomes a thin
delegation layer. No behavioral changes.
